requirements.txt
chainlit>=1.0.0
openai>=1.0.0
huggingface_hub>=0.20.0
pymilvus>=2.3.0
sentence-transformers>=2.2.0
unstructured[all-docs]>=0.10.0
langchain-unstructured>=0.1.0
langchain>=0.1.0
langchain-core>=0.1.0
python-dotenv>=1.0.0
pydantic>=2.0.0
fastapi>=0.100.0
uvicorn>=0.20.0
langchain_nebius>=0.1.0
asyncpg>=0.29.0
ragas>=0.1.0
datasets>=2.0.0
